// src/models/profile.ts
export interface Profile {
  id: string;
  avatar: string;
  birthDate: string;
  contactNumber: string;
  createdAt: string;
  email: string;
  firstName: string;
  lastName: string;
  updatedAt: string;
  gender: string;
  address: string;
  country: string;
  state: string;
  city: string;
  stageName: string;
  serviceType: string;
}